Fix compile warnings.

+++ src/7.in_practice/3.2d_game/0.full_source/game.cpp
@@ -243,7 +243,7 @@ void Game::Render()
         // end rendering to postprocessing framebuffer
         Effects->EndRender();
         // render postprocessing quad
-        Effects->Render(glfwGetTime());
+        Effects->Render(static_cast<float>(glfwGetTime()));
 
         // render text (don't include in postprocessing)
         std::string lives = std::format("Lives : {}", this->Lives);
